+++ client/src/language/ui5/Ui5DefinitionProviders.ts
@@ -1,18 +1,31 @@
-import * as vscode from 'vscode';
 import * as file from '../../helpers/filehandler';
 import * as commands from '../../commands';
-import { TextDocument, Position, CancellationToken, CompletionItem, CompletionList } from 'vscode'
+import {
+    CancellationToken,
+    CompletionItem,
+    CompletionList,
+    Definition,
+    DefinitionProvider,
+    Location,
+    Position,
+    TextDocument,
+    workspace,
+    window,
+    Uri,
+    Range
+} from 'vscode';
 import { channel } from '../../extension'
+import { Storage } from '../xml/XmlDiagnostics'
 
-export class Ui5ViewDefinitionProvider implements vscode.DefinitionProvider {
-    provideDefinition(document: vscode.TextDocument, position: vscode.Position, token: vscode.CancellationToken): Thenable<vscode.Definition> {
+export class Ui5ViewDefinitionProvider implements DefinitionProvider {
+    provideDefinition(document: TextDocument, position: Position, token: CancellationToken): Thenable<Definition> {
         return null
     }
 }
 
-export class Ui5FragmentDefinitionProvider implements vscode.DefinitionProvider {
-    provideDefinition(document: vscode.TextDocument, position: vscode.Position, token: vscode.CancellationToken): Thenable<vscode.Definition> {
-        let line = vscode.window.activeTextEditor.document.lineAt(vscode.window.activeTextEditor.selection.active);
+export class Ui5FragmentDefinitionProvider implements DefinitionProvider {
+    provideDefinition(document: TextDocument, position: Position, token: CancellationToken): Thenable<Definition> {
+        let line = window.activeTextEditor.document.lineAt(window.activeTextEditor.selection.active);
         let tag = line.text.match(/(\w+)Name="(.*?)"/);
 
         if(!tag)
@@ -22,11 +35,11 @@ export class Ui5FragmentDefinitionProvider implements vscode.DefinitionProvider 
         switch (tag[1]) {
             case "view":
                 return file.File.find(new RegExp(tName+"\\.view\\.(xml|json)$")).then((files) => {
-                    return new vscode.Location(vscode.Uri.parse("file:///"+files[0]), new vscode.Position(0,0));
+                    return new Location(Uri.parse("file:///"+files[0]), new Position(0,0));
                 });
             case "fragment":
                 return file.File.find(new RegExp(tName+"\\.fragment\\.(xml|json)$")).then((files) => {
-                    return new vscode.Location(vscode.Uri.parse("file:///"+files[0]), new vscode.Position(0,0));
+                    return new Location(Uri.parse("file:///"+files[0]), new Position(0,0));
                 });
             default:
                 break;
@@ -34,10 +47,10 @@ export class Ui5FragmentDefinitionProvider implements vscode.DefinitionProvider 
     }
 }
 
-export class Ui5ControllerDefinitionProvider implements vscode.DefinitionProvider {
-    provideDefinition(document: vscode.TextDocument, position: vscode.Position, token: vscode.CancellationToken): Thenable<vscode.Definition> {
+export class Ui5ControllerDefinitionProvider implements DefinitionProvider {
+    provideDefinition(document: TextDocument, position: Position, token: CancellationToken): Thenable<Definition> {
         return null;
-        // let line = vscode.window.activeTextEditor.document.lineAt(vscode.window.activeTextEditor.selection.active);
+        // let line = window.activeTextEditor.document.lineAt(window.activeTextEditor.selection.active);
         // let tag = line.text.match(/(\w+)Name="(.*?)"/);
 
         // if(!tag)
@@ -49,24 +62,24 @@ export class Ui5ControllerDefinitionProvider implements vscode.DefinitionProvide
         //         return file.File.find(new RegExp(tName+"\\.controller\\.(js|ts)$")).then((files) => {
         //             // Check typescript (dirty)
         //             let f = files.length>1 ? files[1] : files[0];
-        //             new vscode.Location(vscode.Uri.parse("file:///"+f), new vscode.Position(0,0));
+        //             new Location(Uri.parse("file:///"+f), new Position(0,0));
         //         });
         //     case "view":
         //         return file.File.find(new RegExp(tName+"\\.view\\.(xml|json)$")).then((files) => {
-        //             return new vscode.Location(vscode.Uri.parse("file:///"+files[0]), new vscode.Position(0,0));
+        //             return new Location(Uri.parse("file:///"+files[0]), new Position(0,0));
         //         });
         //     case "fragment":
         //         return file.File.find(new RegExp(tName+"\\.fragment\\.(xml|json)$")).then((files) => {
-        //             return new vscode.Location(vscode.Uri.parse("file:///"+files[0]), new vscode.Position(0,0));
+        //             return new Location(Uri.parse("file:///"+files[0]), new Position(0,0));
         //         });
         //     default:
-        //         let eventhandlertag = vscode.window.activeTextEditor.selection.active;
+        //         let eventhandlertag = window.activeTextEditor.selection.active;
         //         break;
         // }
     }
 }
 
-function tryOpenEventHandler(document: vscode.TextDocument, position: vscode.Position, token: vscode.CancellationToken): Thenable<vscode.Definition> {
+function tryOpenEventHandler(document: TextDocument, position: Position, token: CancellationToken): Thenable<Definition> {
     let rightpart = document.lineAt(position.line).text.substr(position.character).match(/(\w*?)"/)[1];
     if(!rightpart)
         return null;
@@ -78,7 +91,7 @@ function tryOpenEventHandler(document: vscode.TextDocument, position: vscode.Pos
     leftpart = leftpart.substr(leftquotepos);
     let name = leftpart+rightpart;
 
-    let text = vscode.window.activeTextEditor.document.getText();
+    let text = window.activeTextEditor.document.getText();
     let cnameri = text.match(/controllerName="([\w\.]+)"/);
 
     if(!cnameri) {
@@ -89,14 +102,27 @@ function tryOpenEventHandler(document: vscode.TextDocument, position: vscode.Pos
 
     return file.File.find(new RegExp(cname+"\\.controller\\.(js|ts)$")).then((files) => {
                 let f = files.length>1 ? files[1] : files[0];
-                let uri = vscode.Uri.parse("file:///"+f);
-                return vscode.workspace.openTextDocument(uri).then((controllerdoc) => {
+                let uri = Uri.parse("file:///"+f);
+                return workspace.openTextDocument(uri).then((controllerdoc) => {
                     let ccontent = controllerdoc.getText();
 
                     let match = new RegExp(/^(\s*?)/.source+name+/\s*?\(.*?\)/.source, "gm").exec(ccontent);
                     let lineNumber = document.positionAt(match.index + match[1].length).line;
                     let range = controllerdoc.lineAt(lineNumber).range;
-                    return new vscode.Location(uri, range);
+                    return new Location(uri, range);
                 });
             });
+}
+
+export class I18nDfinitionProvider implements DefinitionProvider {
+    public provideDefinition(document: TextDocument, position: Position, token: CancellationToken): Definition | Thenable<Definition> {
+        let i18nlabelregex = new RegExp("\"\s*?{\s*?" + workspace.getConfiguration("ui5ts").get("lang.i18n.modelname") + "\s*?>\s*?(.*?)\s*?}\s*?\"", "g").exec(document.lineAt(position).text);
+        if(i18nlabelregex) {
+            let label = Storage.i18n.labels[i18nlabelregex[1]];
+            return <Location>{
+                range: new Range(label.line, 0, label.line, 1),
+                uri: Storage.i18n.modelfile
+            }
+        }
+    }
 }